requirements.txt
opencv-python
dlib
numpy
tensorflow
spotipy
python-dotenv
webbrowser
